+++ framework/i18n/MessageSource.php
@@ -108,7 +108,8 @@ class MessageSource extends Component
         }
         if (isset($this->_messages[$key][$message]) && $this->_messages[$key][$message] !== '') {
             return $this->_messages[$key][$message];
-        } elseif ($this->hasEventHandlers(self::EVENT_MISSING_TRANSLATION)) {
+        }
+        if ($this->hasEventHandlers(self::EVENT_MISSING_TRANSLATION)) {
             $event = new MissingTranslationEvent([
                 'category' => $category,
                 'message' => $message,