+++ framework/test/FixtureTrait.php
@@ -197,18 +197,18 @@ trait FixtureTrait
         while (($fixture = array_pop($stack)) !== null) {
             if ($fixture instanceof Fixture) {
                 $class = get_class($fixture);
-                $name = isset($aliases[$class]) ? $aliases[$class] : $class;
+                $name = $aliases[$class] ?? $class;
                 unset($instances[$name]);  // unset so that the fixture is added to the last in the next line
                 $instances[$name] = $fixture;
             } else {
                 $class = ltrim($fixture['__class'], '\\');
-                $name = isset($aliases[$class]) ? $aliases[$class] : $class;
+                $name = $aliases[$class] ?? $class;
                 if (!isset($instances[$name])) {
                     $instances[$name] = false;
                     $stack[] = $fixture = Yii::createObject($fixture);
                     foreach ($fixture->depends as $dep) {
                         // need to use the configuration provided in test case
-                        $stack[] = isset($config[$dep]) ? $config[$dep] : ['__class' => $dep];
+                        $stack[] = $config[$dep] ?? ['__class' => $dep];
                     }
                 } elseif ($instances[$name] === false) {
                     throw new InvalidConfigException("A circular dependency is detected for fixture '$class'.");